split up, tests working..
